fix (#30): apply effects to a clone of uiGroup
Applying directly to uiGroup causes glitches during screen recording. In
a similar issue, GNOME devs suggested applying effects to a
Clutter.Clone of uiGroup instead.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2269

Unfortunately, the screen magnifier tool also relies on displaying a
uiGroup clone over the top of it, and the two clones don't get along. We
need to track the magnifier enabling and disabling and attach our effect
accordingly.